refactor: move resolveCssVarColor to utils dir
