+++ src/features/inspector/index.ts
@@ -1,3 +1,4 @@
 export * from './models';
 export * from './viewmodels';
 export * from './views';
+export * from './utils';
