+++ src/features/inspector/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './resolve-css-var-color';
