+++ src/features/inspector/utils/resolve-css-var-color.ts
@@ -1,0 +1,7 @@
+export const resolveCssVarColor = (color: string) => {
+  if (typeof window === 'undefined') return color;
+  const match = color.match(/var\((--[^)]+)\)/);
+  if (!match) return color;
+  const value = getComputedStyle(document.documentElement).getPropertyValue(match[1]).trim();
+  return value || color;
+};
